Unified CLI calls for binaries
