+++ src/storm-cli-utilities/cli.cpp
@@ -1,15 +1,18 @@
 #include "cli.h"
 
+#include "storm-cli-utilities/print.h"
 #include "storm-cli-utilities/resources.h"
-#include "storm-version-info/storm-version.h"
+#include "storm/exceptions/OptionParserException.h"
 #include "storm/io/file.h"
+#include "storm/settings/SettingsManager.h"
+#include "storm/settings/modules/DebugSettings.h"
+#include "storm/settings/modules/GeneralSettings.h"
+#include "storm/settings/modules/ResourceSettings.h"
 #include "storm/utility/SignalHandler.h"
 #include "storm/utility/Stopwatch.h"
 #include "storm/utility/initialize.h"
 #include "storm/utility/macros.h"
 
-#include <boost/algorithm/string/replace.hpp>
-#include <ctime>
 #include <type_traits>
 
 #include "storm-cli-utilities/model-handling.h"
@@ -48,169 +51,6 @@
 
 namespace storm {
 namespace cli {
-
-int64_t process(const int argc, const char** argv) {
-    storm::utility::setUp();
-    storm::cli::printHeader("Storm", argc, argv);
-    storm::settings::initializeAll("Storm", "storm");
-
-    storm::settings::addModule<storm::settings::modules::CounterexampleGeneratorSettings>();
-
-    storm::utility::Stopwatch totalTimer(true);
-    if (!storm::cli::parseOptions(argc, argv)) {
-        return -1;
-    }
-
-    processOptions();
-
-    totalTimer.stop();
-    if (storm::settings::getModule<storm::settings::modules::ResourceSettings>().isPrintTimeAndMemorySet()) {
-        storm::cli::printTimeAndMemoryStatistics(totalTimer.getTimeInMilliseconds());
-    }
-
-    storm::utility::cleanUp();
-    return 0;
-}
-
-std::string shellQuoteSingleIfNecessary(const std::string& arg) {
-    // quote empty argument
-    if (arg.empty()) {
-        return "''";
-    }
-
-    if (arg.find_first_not_of("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_./=") != std::string::npos) {
-        // contains potentially unsafe character, needs quoting
-        if (arg.find('\'') != std::string::npos) {
-            // contains ', we have to replace all ' with '\''
-            std::string escaped(arg);
-            boost::replace_all(escaped, "'", "'\\''");
-            return "'" + escaped + "'";
-        } else {
-            return "'" + arg + "'";
-        }
-    }
-
-    return arg;
-}
-
-void printHeader(std::string const& name, const int argc, const char** argv) {
-    STORM_PRINT(name << " " << storm::StormVersion::shortVersionString() << '\n');
-#ifndef NDEBUG
-    STORM_PRINT("DEBUG BUILD\n");
-#endif
-    // "Compute" the command line argument string with which storm was invoked.
-    std::stringstream commandStream;
-    for (int i = 1; i < argc; ++i) {
-        commandStream << " " << shellQuoteSingleIfNecessary(argv[i]);
-    }
-
-    std::string command = commandStream.str();
-
-    if (!command.empty()) {
-        STORM_PRINT('\n');
-        std::time_t result = std::time(nullptr);
-        STORM_PRINT("Date: " << std::ctime(&result));
-        STORM_PRINT("Command line arguments:" << commandStream.str() << '\n');
-        STORM_PRINT("Current working directory: " << storm::utility::cli::getCurrentWorkingDirectory() << "\n\n");
-    }
-}
-
-void printVersion() {
-    STORM_PRINT(storm::StormVersion::longVersionString() << '\n');
-    STORM_PRINT(storm::StormVersion::buildInfo() << '\n');
-
-    // Print configuration options
-#ifdef STORM_USE_CLN_EA
-    STORM_PRINT("Using CLN numbers for exact arithmetic.\n");
-#else
-    STORM_PRINT("Using GMP numbers for exact arithmetic.\n");
-#endif
-#ifdef STORM_USE_CLN_RF
-    STORM_PRINT("Using CLN numbers for rational functions.\n");
-#else
-    STORM_PRINT("Using GMP numbers for rational functions.\n");
-#endif
-
-    // Print linked dependencies
-#ifdef STORM_HAVE_CARL
-    STORM_PRINT("Linked with CArL v" << STORM_CARL_VERSION << ".\n");
-#endif
-#ifdef STORM_HAVE_GLPK
-    STORM_PRINT("Linked with GNU Linear Programming Kit v" << GLP_MAJOR_VERSION << "." << GLP_MINOR_VERSION << ".\n");
-#endif
-#ifdef STORM_HAVE_GUROBI
-    STORM_PRINT("Linked with Gurobi Optimizer v" << GRB_VERSION_MAJOR << "." << GRB_VERSION_MINOR << "." << GRB_VERSION_TECHNICAL << ".\n");
-#endif
-#ifdef STORM_HAVE_HYPRO
-    STORM_PRINT("Linked with HyPro v" << STORM_HYPRO_VERSION << ".\n");
-#endif
-#ifdef STORM_HAVE_INTELTBB
-    STORM_PRINT("Linked with Intel Threading Building Blocks v" << TBB_VERSION_MAJOR << "." << TBB_VERSION_MINOR << " (Interface version "
-                                                                << TBB_INTERFACE_VERSION << ").\n");
-#endif
-#ifdef STORM_HAVE_MSAT
-    char* msatVersion = msat_get_version();
-    STORM_PRINT("Linked with " << msatVersion << ".\n");
-    msat_free(msatVersion);
-#endif
-#ifdef STORM_HAVE_SOPLEX
-    STORM_PRINT("Linked with Soplex: " << soplex::printVersion() << ".\n");
-#endif
-#ifdef STORM_HAVE_SMTRAT
-    STORM_PRINT("Linked with SMT-RAT v" << SMTRAT_VERSION << ".\n");
-#endif
-#ifdef STORM_HAVE_SPOT
-    STORM_PRINT("Linked with Spot v" << spot::version() << ".\n");
-#endif
-#ifdef STORM_HAVE_XERCES
-    STORM_PRINT("Linked with Xerces-C v" << gXercesMajVersion << "." << gXercesMinVersion << "." << gXercesRevision << ".\n");
-#endif
-#ifdef STORM_HAVE_Z3
-    unsigned int z3Major, z3Minor, z3BuildNumber, z3RevisionNumber;
-    Z3_get_version(&z3Major, &z3Minor, &z3BuildNumber, &z3RevisionNumber);
-#ifdef STORM_HAVE_Z3_OPTIMIZE
-    STORM_PRINT("Linked with Z3 Theorem Prover v" << z3Major << "." << z3Minor << " Build " << z3BuildNumber << " Rev " << z3RevisionNumber
-                                                  << " (with optimization features).\n");
-#else
-    STORM_PRINT("Linked with Z3 Theorem Prover v" << z3Major << "." << z3Minor << " Build " << z3BuildNumber << " Rev " << z3RevisionNumber << ".\n");
-#endif
-#endif
-
-#ifdef STORM_HAVE_CUDA
-    int deviceCount = 0;
-    cudaError_t error_id = cudaGetDeviceCount(&deviceCount);
-
-    if (error_id == cudaSuccess) {
-        STORM_PRINT("Compiled with CUDA support, ");
-        // This function call returns 0 if there are no CUDA capable devices.
-        if (deviceCount == 0) {
-            STORM_PRINT("but there are no available device(s) that support CUDA.\n");
-        } else {
-            STORM_PRINT("detected " << deviceCount << " CUDA capable device(s):\n");
-        }
-
-        int dev, driverVersion = 0, runtimeVersion = 0;
-
-        for (dev = 0; dev < deviceCount; ++dev) {
-            cudaSetDevice(dev);
-            cudaDeviceProp deviceProp;
-            cudaGetDeviceProperties(&deviceProp, dev);
-
-            STORM_PRINT("CUDA device " << dev << ": \"" << deviceProp.name << "\"\n");
-
-            // Console log
-            cudaDriverGetVersion(&driverVersion);
-            cudaRuntimeGetVersion(&runtimeVersion);
-            STORM_PRINT("  CUDA Driver Version / Runtime Version          " << driverVersion / 1000 << "." << (driverVersion % 100) / 10 << " / "
-                                                                            << runtimeVersion / 1000 << "." << (runtimeVersion % 100) / 10 << '\n');
-            STORM_PRINT("  CUDA Capability Major/Minor version number:    " << deviceProp.major << "." << deviceProp.minor << '\n');
-        }
-        STORM_PRINT('\n');
-    } else {
-        STORM_PRINT("Compiled with CUDA support, but an error occurred trying to find CUDA devices.\n");
-    }
-#endif
-}
 
 bool parseOptions(const int argc, const char* argv[]) {
     try {
@@ -272,73 +112,37 @@ void setLogLevel() {
     setFileLogging();
 }
 
-void setUrgentOptions() {
+int process(std::string const& name, std::string const& executableName, std::function<void(std::string const&, std::string const&)> initSettingsFunc,
+            std::function<void(void)> processOptionsFunc, const int argc, const char** argv) {
+    storm::utility::setUp();
+    storm::cli::printHeader(name, argc, argv);
+
+    // Initialize settings
+    initSettingsFunc(name, executableName);
+
+    storm::utility::Stopwatch totalTimer(true);
+    if (!parseOptions(argc, argv)) {
+        return -1;
+    }
+
+    // Start by setting some urgent options (log levels, resources, etc.)
     setResourceLimits();
     setLogLevel();
     setFileLogging();
     // Set output precision
     storm::utility::setOutputDigitsFromGeneralPrecision(storm::settings::getModule<storm::settings::modules::GeneralSettings>().getPrecision());
-}
 
-void processOptions() {
-    // Start by setting some urgent options (log levels, resources, etc.)
-    setUrgentOptions();
+    // Process options and start computations
+    processOptionsFunc();
 
-    // Parse symbolic input (PRISM, JANI, properties, etc.)
-    SymbolicInput symbolicInput = parseSymbolicInput();
-
-    // Obtain settings for model processing
-    ModelProcessingInformation mpi;
-
-    // Preprocess the symbolic input
-    std::tie(symbolicInput, mpi) = preprocessSymbolicInput(symbolicInput);
-
-    STORM_LOG_WARN_COND(mpi.isCompatible,
-                        "The model checking query does not seem to be supported for the selected engine. Storm will try to solve the query, but you will most "
-                        "likely get an error for at least one of the provided properties.");
-
-    // Export symbolic input (if requested)
-    exportSymbolicInput(symbolicInput);
-
-#ifdef STORM_HAVE_CARL
-    switch (mpi.verificationValueType) {
-        case ModelProcessingInformation::ValueType::Parametric:
-            processInputWithValueType<storm::RationalFunction>(symbolicInput, mpi);
-            break;
-        case ModelProcessingInformation::ValueType::Exact:
-            processInputWithValueType<storm::RationalNumber>(symbolicInput, mpi);
-            break;
-        case ModelProcessingInformation::ValueType::FinitePrecision:
-            processInputWithValueType<double>(symbolicInput, mpi);
-            break;
+    totalTimer.stop();
+    if (storm::settings::getModule<storm::settings::modules::ResourceSettings>().isPrintTimeAndMemorySet()) {
+        storm::cli::printTimeAndMemoryStatistics(totalTimer.getTimeInMilliseconds());
     }
-#else
-    STORM_LOG_THROW(mpi.verificationValueType == ModelProcessingInformation::ValueType::FinitePrecision, storm::exceptions::NotSupportedException,
-                    "No exact numbers or parameters are supported in this build.");
-    processInputWithValueType<double>(symbolicInput, mpi);
-#endif
-}
 
-void printTimeAndMemoryStatistics(uint64_t wallclockMilliseconds) {
-    struct rusage ru;
-    getrusage(RUSAGE_SELF, &ru);
-
-    std::cout << "\nPerformance statistics:\n";
-#ifdef MACOS
-    // For Mac OS, this is returned in bytes.
-    uint64_t maximumResidentSizeInMegabytes = ru.ru_maxrss / 1024 / 1024;
-#endif
-#ifdef LINUX
-    // For Linux, this is returned in kilobytes.
-    uint64_t maximumResidentSizeInMegabytes = ru.ru_maxrss / 1024;
-#endif
-    std::cout << "  * peak memory usage: " << maximumResidentSizeInMegabytes << "MB\n";
-    char oldFillChar = std::cout.fill('0');
-    std::cout << "  * CPU time: " << ru.ru_utime.tv_sec << "." << std::setw(3) << ru.ru_utime.tv_usec / 1000 << "s\n";
-    if (wallclockMilliseconds != 0) {
-        std::cout << "  * wallclock time: " << (wallclockMilliseconds / 1000) << "." << std::setw(3) << (wallclockMilliseconds % 1000) << "s\n";
-    }
-    std::cout.fill(oldFillChar);
+    // All operations have been performed, so we clean up everything and terminate.
+    storm::utility::cleanUp();
+    return 0;
 }
 
 }  // namespace cli
